Remove ethereum composition
